restructure tests dir - move current tests to tests/base_tests to separate testing base package from testing extras
