+++ tests/base_tests/classification/test_coverages_calculation.py
@@ -8,14 +8,13 @@ import pandas as pd
 from decision_rules.classification.rule import ClassificationRule
 from decision_rules.classification.ruleset import ClassificationRuleSet
 from decision_rules.serialization import JSONSerializer
+from tests.loaders import load_resources_path
 
 
 class TestClassificationCoveragesCalculation(unittest.TestCase):
 
     def _prepare_test_rule(self) -> ClassificationRule:
-        dir_path = os.path.dirname(os.path.realpath(__file__))
-        ruleset_file_path: str = os.path.join(
-            dir_path, '..', 'resources', 'iris_ruleset.json')
+        ruleset_file_path: str = os.path.join(load_resources_path(), 'iris_ruleset.json')
 
         with open(ruleset_file_path, 'r', encoding='utf-8') as file:
             return JSONSerializer.deserialize(
@@ -24,9 +23,7 @@ class TestClassificationCoveragesCalculation(unittest.TestCase):
             )
 
     def _prepare_test_dataset(self) -> tuple[pd.DataFrame, pd.Series]:
-        dir_path = os.path.dirname(os.path.realpath(__file__))
-        dataset_path: str = os.path.join(
-            dir_path, '..', 'resources', 'iris.csv')
+        dataset_path: str = os.path.join(load_resources_path(), 'iris.csv')
 
         df = pd.read_csv(dataset_path)
         X = df.drop('class', axis=1)
