+++ tests/base_tests/core/test_abstract_condition.py
