+++ tests/base_tests/core/test_measures.py
