+++ tests/base_tests/core/test_nominal_condition.py
