+++ tests/base_tests/core/test_numerical_conditions.py
