+++ tests/base_tests/core/test_rules_metrics.py
