+++ tests/base_tests/serialization/core/test_condition_serializer.py
