+++ tests/base_tests/serialization/regression/test_rule_serializer.py
