+++ tests/base_tests/serialization/survival/test_ruleset_serializer.py
